Add Github Actions CI
We want to have a release build, a code coverage build, a formatting
check, and a linting run with clippy.

The code coverage information will be uploaded to coveralls.io. I'm not
sure if that is the best service, but it was readily available and
documented in the `grcov` action.